spott_backend/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
pymongo==4.7.3
flask-jwt-extended==4.6.0
bcrypt==4.1.3
python-dotenv==1.0.1
Flask-CORS==4.0.1
Pillow==10.4.0
python-slugify==8.0.4
email-validator==2.2.0
Flask-Limiter==3.7.0
gunicorn==22.0.0
marshmallow==3.21.3
qrcode==7.4.2
Werkzeug==3.0.3
python-dateutil==2.9.0.post0
